LayoutTests/fast/notifications/notifications-request-permission-optional-expected.txt

PASS window.webkitNotifications is defined.
PASS window.webkitNotifications.requestPermission() did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
